requirements.txt
crewai
google-api-python-client
google-auth-oauthlib
slack-sdk
python-dotenv
python-dateutil